+++ R/woebin.R
@@ -1,27 +1,83 @@
-# [tree-like]
-# [chimerge](http://blog.csdn.net/qunxingvip/article/details/50449376)
+# woebin woebin_plot woebin_ply woebin_adj
 
+# converting vector (breaks & special_values) to dataframe
+split_vec_todf = function(vec) {
+  value = . = bin_chr = V1 = NULL
 
-# required in woebin2 # return binning if provide breaks
+  if (!is.null(vec)) data.table(
+    value=vec, bin_chr=vec
+  )[, rowid := .I
+  ][, strsplit(as.character(value), "%,%", fixed=TRUE), by = .(rowid, bin_chr)
+  ][, .(rowid, bin_chr, value = ifelse(V1=="missing", NA, as.character(V1)) )]
+}
+# add missing to spl_val if there is na in dtm$value and
+# missing is not specified in breaks and spl_val
+add_missing_spl_val = function(dtm, breaks, spl_val) {
+  value = NULL
+
+  if (dtm[,any(is.na(value))]) {
+    no_missing = !any(grepl('missing', c(breaks, spl_val)))
+    if (no_missing) {
+      spl_val = c('missing',spl_val)
+    }
+  }
+  return(spl_val)
+}
+# split dtm into bin_sv and dtm (without speical_values)
+dtm_binning_sv = function(dtm, breaks, spl_val) {
+  binning_sv = value = . = y = variable = good = bad = bin = NULL
+  # spl_val
+  spl_val = add_missing_spl_val(dtm, breaks, spl_val)
+  if (!is.null(spl_val)) {
+    # special_values from vector to dataframe
+    sv_df = split_vec_todf(spl_val)
+
+    # dtm_sv & dtm
+    dtm_sv = setDT(dtm)[value %in% sv_df$value]
+    dtm = setDT(dtm)[!(value %in% sv_df$value)]
+
+    # binning_sv
+    binning_sv = merge(
+      dtm_sv[, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , by = value][,value:=as.character(value)],
+      sv_df[,value:=as.character(value)],
+      all.x = TRUE, by='value'
+    )[, value:=ifelse(is.na(value), "missing", as.character(value))
+    ][, .(bin=paste0(value,collapse="%,%"), good=sum(good), bad=sum(bad), variable=unique(variable)), by=rowid
+    ][, .(variable, bin, good, bad)]
+  }
+
+  return(list(binning_sv=binning_sv, dtm=dtm))
+}
+
+# required in woebin2 # return binning if breaks provided
 #' @import data.table
-woebin2_breaks = function(dtm, breaks) {
+woebin2_breaks = function(dtm, breaks, spl_val) {
   # global variables or functions
-  value = bin = . = y = variable = bad = good = V1 = badprob = NULL
+  value = bin = . = y = variable = bad = good = V1 = badprob = bksv_list = bin_chr = NULL
 
-  dtm = setDT(dtm)
+  # breaks from vector to dataframe
+  bk_df = split_vec_todf(breaks)
+
+  # dtm $ binning_sv
+  dtm_binsv_list = dtm_binning_sv(dtm, breaks, spl_val)
+  dtm = dtm_binsv_list$dtm
+  binning_sv = dtm_binsv_list$binning_sv
+  if (dtm[,.N] == 0) return(list(binning_sv=binning_sv, binning=NULL))
+
+
+  # binning
   if (is.numeric(dtm[,value])) {
-    # numeric variable ------
-    bstbrks = c(-Inf, setdiff(unique(breaks), c(NA, Inf, -Inf)), Inf)
+    bstbrks = c(-Inf, setdiff(unique(bk_df$value), c(NA, Inf, -Inf)), Inf)
 
     binning = dtm[
       , bin := cut(value, bstbrks, right = FALSE, dig.lab = 10, ordered_result = FALSE)]
 
-    # break points from bin
+    # check empty bins
+    ## break points from bin
     breaks_list = lapply(
       list(left="\\1", right="\\2"),
-      function(x) setdiff(sub("^\\[(.*),(.*)\\)", x, unique(binning$bin)), c(" Inf", "Inf", " -Inf", "-Inf")) )
-
-    # if there are empty bins
+      function(x) setdiff(sub("^\\[(.*), *(.*)\\)", x, unique(binning$bin)), c("Inf","-Inf")) )
+    ## if there are empty bins
     if (!setequal(breaks_list$left, breaks_list$right)) {
       bstbrks = unique(c(-Inf, unique(breaks_list$right), Inf))
       binning = dtm[
@@ -31,29 +87,53 @@ woebin2_breaks = function(dtm, breaks) {
     }
 
 
-    binning = binning[, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , keyby = bin
+    binning = binning[, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , by = .(bin)
       ][order(bin)]
 
+    # merge binning with bk_df
+    if (bk_df[is.na(value),.N] == 1) {
+      binning = merge(
+        binning[, value:=sub("^\\[(.*), *(.*)\\)","\\2",bin)],
+        bk_df,
+        all.x = TRUE, by="value"
+      )[order(rowid,value)][, bin:=ifelse(is.na(bin), "missing", as.character(bin))
+      ][, .(bin=paste0(bin,collapse="%,%"), good=sum(good), bad=sum(bad), variable=unique(variable)), by=rowid
+      ][order(rowid)]
+    }
+
   } else if (is.factor(dtm[,value]) || is.character(dtm[,value])) {
-    # other variable ------
     dtm = dtm[,value := as.character(value)]
 
+    # merge binning with bk_df
     binning = merge(
-      data.table(rowid=1:length(breaks), bin = breaks, value = breaks )[, strsplit(as.character(value), "%,%", fixed=TRUE), by = c("rowid","bin") ][, .(rowid, bin, value = V1)],
-      dtm, all.y = TRUE
-    )[, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , keyby = c("rowid","bin")]
+      dtm, bk_df[,bin:=bin_chr], all.x = TRUE
+    )[order(rowid, bin)][, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , by = .(rowid, bin)]
 
   }
 
-  return(binning)
+
+  # # remove rowid column in binning dataframe
+  binning = binning[,rowid:=1][,rowid:=NULL]
+  # # bind binning_sv and binning
+  # if (setDT(binning_sv)[,.N] > 0) binning = rbind(binning_sv, binning)
+
+  return(list(binning_sv=binning_sv, binning=binning))
 }
 
 # required in woebin2 # return initial binning
-woebin2_init_bin = function(dtm, min_perc_total) {
+woebin2_init_bin = function(dtm, min_perc_fine_bin, breaks, spl_val) {
   # global variables or functions
-  value = bin = . = y = variable = bad = good = brkp = badprob = count = merge_tolead = brkp2 = NULL
+  value = bin = . = y = variable = bad = good = brkp = badprob = count = merge_tolead = brkp2 = count_lag = count_lead = NULL
 
-  dtm = setDT(dtm)
+
+  # dtm $ binning_sv
+  dtm_binsv_list = dtm_binning_sv(dtm, breaks, spl_val)
+  dtm = dtm_binsv_list$dtm
+  binning_sv = dtm_binsv_list$binning_sv
+  if (is.null(dtm) || dtm[,.N]==0) return(list(binning_sv=binning_sv, initial_binning=NULL))
+
+
+  # binning
   if (is.numeric(dtm[,value])) {
     # numeric variable ------
     xvalue = dtm[, value]
@@ -64,11 +144,11 @@ woebin2_init_bin = function(dtm, min_perc_total) {
     if (iqr == 0) {
       xvalue_rm_outlier = xvalue
     } else {
-      xvalue_rm_outlier = xvalue[which(xvalue > iq[2]-3*iqr & xvalue <= iq[4]+3*iqr)]
+      xvalue_rm_outlier = xvalue[which(xvalue >= iq[2]-3*iqr & xvalue <= iq[4]+3*iqr)]
     }
 
     # number of initial binning
-    n = trunc(1/min_perc_total)
+    n = trunc(1/min_perc_fine_bin)
     len_uniq_x = length(setdiff(unique(xvalue_rm_outlier), c(NA,Inf,-Inf)))
     if (len_uniq_x < n) n = len_uniq_x
 
@@ -79,13 +159,13 @@ woebin2_init_bin = function(dtm, min_perc_total) {
       brk = pretty(xvalue_rm_outlier, n)
     }
     brk = sort(brk)
-    brk = unique(c(-Inf, brk[2:(length(brk)-1)], Inf))
+    brk = unique(c(-Inf, brk[2:length(brk)], Inf))
     if (anyNA(xvalue)) brk = c(brk, NA)
 
     # initial binning datatable
     init_bin = dtm[
       , bin := cut(value, brk, right = FALSE, dig.lab = 10, ordered_result = FALSE)
-    ][, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , keyby = bin
+    ][, .(good = sum(y==0), bad = sum(y==1), variable=unique(variable)) , by = bin
     ][order(bin)
     ][, `:=`(brkp = as.numeric( sub("^\\[(.*),.+", "\\1", bin)), badprob = bad/(good+bad))
     ][, .(variable, bin, brkp, good, bad, badprob)]
@@ -123,42 +203,42 @@ woebin2_init_bin = function(dtm, min_perc_total) {
     # brkp needs to be removed if good==0 or bad==0
     rm_brkp = init_bin[!is.na(brkp)][
       ,count := good+bad
-    ][, merge_tolead := shift(count,type="lag", fill=nrow(dtm)+1) > shift(count,type="lead", fill=nrow(dtm)+1)
+    ][,`:=`(
+      count_lag=shift(count,type="lag", fill=nrow(dtm)+1),
+      count_lead=shift(count,type="lead", fill=nrow(dtm)+1)
+   )][, merge_tolead := count_lag > count_lead
     ][good == 0 | bad == 0][count == min(count)]
 
     # set brkp to lead's or lag's
-    if (rm_brkp[1,merge_tolead]) {
-      init_bin = init_bin[
-        ,brkp2 := shift(brkp,type="lead")
-      ][brkp == rm_brkp[1,brkp], brkp := brkp2]
-
-    } else {
-
-      init_bin = init_bin[
-        ,brkp2 := shift(brkp,type="lag")
-      ][brkp == rm_brkp[1,brkp], brkp := brkp2]
-    }
+    shift_type = ifelse(rm_brkp[1,merge_tolead], 'lead', 'lag')
+    init_bin = init_bin[
+      ,brkp2 := shift(brkp,type=shift_type)
+    ][brkp == rm_brkp[1,brkp], brkp := brkp2]
 
     # groupby brkp
     init_bin = init_bin[
-      ,.(variable=unique(variable), bin=paste0(bin, collapse = "%,%"), brkp=unique(brkp), good=sum(good), bad=sum(bad)), by=brkp
+      ,.(variable=unique(variable), bin=paste0(bin, collapse = "%,%"), good=sum(good), bad=sum(bad)), by=brkp
     ][, badprob:=bad/(good+bad)]
   }
 
+  # format bin
   if (is.numeric(dtm[,value])) {
-    init_bin = init_bin[grepl("%,%",bin), bin := sub("^(\\[.+?,).+,(.+?\\))$", "\\1\\2", bin)]
+    init_bin = init_bin[
+      grepl("%,%",bin), bin := sub("^(\\[.+?,).+,(.+?\\))$", "\\1\\2", bin)
+    ][, brkp := as.numeric(sub("^\\[(.*),.+", "\\1", bin))]
   }
 
-  return(init_bin)
+  return(list(binning_sv=binning_sv, initial_binning=init_bin))
 }
 
 # required in woebin2_tree # add 1 best break for tree-like binning
-woebin2_tree_1bst = function(dtm, initial_binning, min_perc_total, bestbreaks=NULL) {
+woebin2_tree_add_1brkp = function(dtm, initial_binning, min_perc_coarse_bin, bestbreaks=NULL) {
   # global variables or functions
   brkp = patterns = . = good = bad = variable = count_distr = value = min_count_distr = bstbin = min_count_distr = total_iv = bstbin = brkp = bin = NULL
 
+
   # total_iv for all best breaks
-  total_iv_all_breaks = function(initial_binning, bestbreaks) {
+  total_iv_all_breaks = function(initial_binning, bestbreaks, dtm_rows) {
     # best breaks set
     breaks_set = setdiff( initial_binning[,brkp], c(bestbreaks, -Inf, Inf, NA) )
 
@@ -177,21 +257,14 @@ woebin2_tree_1bst = function(dtm, initial_binning, min_perc_total, bestbreaks=NU
     total_iv_all_brks = melt(
       init_bin_all_breaks, id = c("variable", "good", "bad"), variable.name = "bstbin", measure = patterns("bstbin.+")
     )[, .(good = sum(good), bad = sum(bad), variable = unique(variable))
-      , keyby=c("bstbin", "value")
-    ][, count_distr := (good+bad)/sum(good+bad), by="bstbin"
-    ][!is.na(value), min_count_distr := min(count_distr), by="bstbin"
-    ][, .(total_iv = iv_01(good, bad), variable = unique(variable), min_count_distr = min(min_count_distr,na.rm=TRUE)), keyby="bstbin"
+      , by=.(bstbin, value)
+    ][, count_distr := (good+bad)/dtm_rows, by=bstbin
+    ][!is.na(value), min_count_distr := min(count_distr), by=bstbin
+    ][, .(total_iv = iv_01(good, bad), variable = unique(variable), min_count_distr = min(min_count_distr,na.rm=TRUE)), by=bstbin
     ][, bstbin := as.numeric(sub("bstbin(.+)", "\\1", bstbin))][]
 
     return(total_iv_all_brks)
   }
-  total_iv_all_brks = total_iv_all_breaks(initial_binning, bestbreaks)
-
-  # bestbreaks: total_iv==max(total_iv) & min(count_distr)>=min_perc_total*2.5(0.05)
-  bstbrk_max_iv = total_iv_all_brks[min_count_distr >= min_perc_total*2][total_iv==max(total_iv), bstbin]
-  # add 1best break to bestbreaks
-  bestbreaks = unique(c(bestbreaks, bstbrk_max_iv[1]))
-
   # binning add 1best break
   binning_add_1bst = function(initial_binning, bestbreaks) {
     value = bstbin = . = good = bad = variable = woe = bin_iv = total_iv = bstbrkp = badprob = NULL # no visible binding for global variable
@@ -199,7 +272,7 @@ woebin2_tree_1bst = function(dtm, initial_binning, min_perc_total, bestbreaks=NU
     if ( is.numeric(dtm[,value]) ) {
       binning_1bst_brk = initial_binning[
         , bstbin := cut(brkp, c(-Inf, bestbreaks, Inf), right = FALSE, dig.lab = 10, ordered_result = FALSE)
-        ][, .(variable=unique(variable), bin=unique(bstbin), good = sum(good), bad = sum(bad)) , keyby = bstbin
+        ][, .(variable=unique(variable), bin=unique(bstbin), good = sum(good), bad = sum(bad)) , by = bstbin
           ]
 
     } else if (is.logical(dtm[,value]) || is.factor(dtm[,value]) || is.character(dtm[,value]) ) {
@@ -208,64 +281,185 @@ woebin2_tree_1bst = function(dtm, initial_binning, min_perc_total, bestbreaks=NU
 
       binning_1bst_brk = initial_binning[
         , bstbin := cut(brkp, c(-Inf, bestbreaks, Inf), right = FALSE,dig.lab = 10, ordered_result = FALSE)
-        ][, .(variable=unique(variable), bin = paste0(bin, collapse = "%,%"), good = sum(good), bad = sum(bad)), keyby = bstbin ]
+        ][, .(variable=unique(variable), bin = paste0(bin, collapse = "%,%"), good = sum(good), bad = sum(bad)), by = bstbin ]
     }
 
     binning_1bst_brk = binning_1bst_brk[
       order(bstbin)
-    ][, total_iv := iv_01(good, bad)
-    ][, bstbrkp := as.numeric( sub("^\\[(.*),.+", "\\1", bstbin) )
-    ][, .(variable, bin, bstbin, bstbrkp, good, bad, total_iv)]
+      ][, total_iv := iv_01(good, bad)
+        ][, bstbrkp := as.numeric( sub("^\\[(.*),.+", "\\1", bstbin) )
+          ][, .(variable, bin, bstbin, bstbrkp, good, bad, total_iv)]
 
     return(binning_1bst_brk)
   }
+
+
+  # adding 1 best breakpoint
+  dtm_rows = nrow(dtm)
+  total_iv_all_brks = total_iv_all_breaks(initial_binning, bestbreaks, dtm_rows)
+
+  # bestbreaks: total_iv == max(total_iv) & min(count_distr) >= min_perc_coarse_bin
+  bstbrk_max_iv = total_iv_all_brks[min_count_distr >= min_perc_coarse_bin][total_iv==max(total_iv)][, bstbin]
+  # add 1best break to bestbreaks
+  bestbreaks = unique(c(bestbreaks, bstbrk_max_iv[1]))
   bin_add_1bst = binning_add_1bst(initial_binning, bestbreaks)
 
   return(bin_add_1bst)
 }
+
 # required in woebin2 # return tree-like binning
-woebin2_tree = function(dtm, initial_binning, min_perc_total=0.02, stop_limit=0.1, max_bin_num=6) {
+woebin2_tree = function(dtm, min_perc_fine_bin=0.02, min_perc_coarse_bin=0.05, stop_limit=0.1, max_num_bin=8, breaks=NULL, spl_val=NULL) {
   # global variables or functions
   brkp = bstbrkp = total_iv = binning_tree = NULL
 
+  # initial binning
+  bin_list = woebin2_init_bin(dtm, min_perc_fine_bin=min_perc_fine_bin, breaks=breaks, spl_val=spl_val)
+  initial_binning = bin_list$initial_binning
+  binning_sv = bin_list$binning_sv
+
   # initialize parameters
-  len_brks = initial_binning[!is.na(brkp), .N] ## length all breaks
+  ## length all breaks
+  len_brks = initial_binning[!is.na(brkp), .N]
+  ## param
   bestbreaks = NULL ## best breaks
   IVt1 = IVt2 = 1e-10
   IVchg = 1 ## IV gain ratio
   step_num = 1
 
   # best breaks from three to n+1 bins
-  while ( IVchg >= stop_limit & step_num+1 <= min(max_bin_num, len_brks) ) {
-    binning_tree = woebin2_tree_1bst(dtm, initial_binning, min_perc_total, bestbreaks)
+  while ( (IVchg >= stop_limit) & (step_num+1 <= min(max_num_bin, len_brks)) ) {
+    binning_tree = woebin2_tree_add_1brkp(dtm, initial_binning, min_perc_coarse_bin, bestbreaks)
     # print(binning_tree)
 
-    # best breaks
+    # update parameters
+    ## best breaks
     bestbreaks = binning_tree[bstbrkp != -Inf & !is.na(bstbrkp), bstbrkp]
-    # print(bestbreaks)
-
-    # information value
+    ## information value
     IVt2 = binning_tree[1, total_iv]
     IVchg = IVt2/IVt1-1 ## ratio gain
     IVt1 = IVt2
     # print(IVchg)
-
     step_num = step_num + 1
   }
 
   if (is.null(binning_tree)) binning_tree = initial_binning
 
-  return(binning_tree)
+  return(list(binning_sv=binning_sv, binning=binning_tree))
+  # return(binning_tree)
 }
 # examples
-# system.time( initial_binning <- woebin2_init_bin(dtm, min_perc_total) )
-# system.time( woebin2_tree_1bst(dtm, initial_binning, min_perc_total) )
-# system.time( woebin2_tree(dtm, initial_binning, min_perc_total) )
+# system.time( binning_list <- woebin2_init_bin(dtm, min_perc_fine_bin=0.02, breaks =NULL, spl_val=NULL) )
+# initial_binning=binning_list$initial_binning
+# binning_sv = binning_list$binning_sv
+# system.time( woebin2_tree_add_1brkp(dtm, initial_binning, min_perc_coarse_bin=0.05) )
+# system.time( woebin2_tree(dtm, initial_binning, min_perc_coarse_bin=0.05) )
+
+# required in woebin2 # return chimerge binning
+#' @importFrom stats qchisq
+woebin2_chimerge = function(dtm, min_perc_fine_bin=0.02, min_perc_coarse_bin=0.05, stop_limit=0.1, max_num_bin=8, breaks=NULL, spl_val=NULL) {
+  .= a= a_colsum= a_lag= a_lag_rowsum= a_rowsum= a_sum= bad= bin= brkp= brkp2= chisq= count= count_distr= e= e_lag= chisq_lead= good= goodbad= merge_tolead =value= variable= NULL
+
+  # [chimerge](http://blog.csdn.net/qunxingvip/article/details/50449376)
+  # [ChiMerge:Discretization of numeric attributs](http://www.aaai.org/Papers/AAAI/1992/AAAI92-019.pdf)
+
+  # chisq = function(a11, a12, a21, a22) {
+  #   A = list(a1 = c(a11, a12), a2 = c(a21, a22))
+  #   Adf = do.call(rbind, A)
+  #
+  #   Edf =
+  #     matrix(rowSums(Adf), ncol = 1) %*%
+  #     matrix(colSums(Adf), nrow = 1) /
+  #     sum(Adf)
+  #
+  #   sum((Adf-Edf)^2/Edf)
+  # }
+  # initial binning
+  bin_list = woebin2_init_bin(dtm, min_perc_fine_bin=min_perc_fine_bin, breaks=breaks, spl_val=spl_val)
+
+  initial_binning = bin_list$initial_binning
+  binning_sv = bin_list$binning_sv
+
+  # function to create a chisq column in initial_binning
+  add_chisq = function(initial_binning) {
+  chisq_df = melt(initial_binning[!is.na(brkp)], id.vars = c("brkp", "variable", "bin"), measure.vars = c("good", "bad"), variable.name = "goodbad", value.name = "a"
+  )[order(brkp)
+  ][, a_lag := shift(a, type="lag"), by=.(goodbad)
+  ][, `:=`(
+    a_rowsum = sum(a),
+    a_lag_rowsum = sum(a_lag),
+    a_colsum = a+a_lag,
+    a_sum = sum(a+a_lag)), by=brkp
+  ][, `:=`(
+    e = a_rowsum*a_colsum/a_sum,
+    e_lag = a_lag_rowsum*a_colsum/a_sum
+  )][, .(chisq=sum((a-e)^2/e + (a_lag-e_lag)^2/e_lag)), by=brkp]
+
+  return(merge(initial_binning[,count:=good+bad], chisq_df, all.x = TRUE))
+  }
+
+  # dtm_rows
+  dtm_rows = nrow(dtm)
+  # chisq limit
+  chisq_limit = qchisq(1-stop_limit,1)
+  # binning with chisq column
+  binning_chisq = add_chisq(initial_binning)
+
+  # param
+  bin_chisq_min = binning_chisq[, min(chisq, na.rm = TRUE)]
+  bin_count_distr_min = binning_chisq[!is.na(brkp), min((good+bad)/dtm_rows)]
+  bin_nrow = binning_chisq[,.N]
+  # remove brkp if chisq < chisq_limit
+  while (
+    bin_chisq_min < chisq_limit ||
+    bin_count_distr_min < min_perc_coarse_bin ||
+    bin_nrow > max_num_bin) {
+    # brkp needs to be removed
+    if (bin_chisq_min < chisq_limit) {
+      rm_brkp = binning_chisq[, merge_tolead := FALSE][order(chisq, count)][1,]
+
+    } else if (bin_count_distr_min < min_perc_coarse_bin) {
+      rm_brkp = binning_chisq[,`:=`(
+        count_distr = count/sum(count),
+        chisq_lead = shift(chisq, type = "lead", fill = Inf)
+      )][,merge_tolead := ifelse(is.na(chisq), TRUE, chisq > chisq_lead)
+       ][!is.na(brkp)][order(count_distr)][1,]
+
+    } else if (bin_nrow > max_num_bin) {
+      rm_brkp = binning_chisq[, merge_tolead := FALSE][order(chisq, count)][1,]
+
+    }
+
+
+    # groupby brkp
+    shift_type = ifelse(rm_brkp[1,merge_tolead], 'lead', 'lag')
+    binning_chisq = binning_chisq[
+      ,brkp2 := shift(brkp,type=shift_type)
+    ][brkp == rm_brkp[1,brkp], brkp := brkp2
+    ][,.(variable=unique(variable), bin=paste0(bin, collapse = "%,%"), good=sum(good), bad=sum(bad)), by=brkp
+    ]#[, badprob:=bad/(good+bad)]
+
+    # update
+    ## add chisq to new binning dataframe
+    binning_chisq = add_chisq(binning_chisq)
+    ## param
+    bin_chisq_min = binning_chisq[, min(chisq, na.rm = TRUE)]
+    bin_count_distr_min = binning_chisq[!is.na(brkp), min((good+bad)/dtm_rows)]
+    bin_nrow = binning_chisq[,.N]
+  }
+
+  # format bin # remove (.+\\)%,%\\[.+,)
+  if (is.numeric(dtm[,value])) {
+    binning_chisq = binning_chisq[grepl("%,%",bin), bin := sub("^(\\[.+?,).+,(.+?\\))$", "\\1\\2", bin)]
+  }
+
+  return(list(binning_sv=binning_sv, binning=binning_chisq))
+  # return(binning_chisq)
+}
 
 # required in woebin2 # # format binning output
 binning_format = function(binning) {
   # global variables or functions
-  . = bad = badprob = bin = bin_iv = good = total_iv = variable = woe = NULL
+  . = bad = badprob = bin = bin_iv = good = total_iv = variable = woe = is_sv = NULL
   # required columns in input binning: variable, bin, good, bad
   binning = binning[
     , badprob:=bad/(good+bad)
@@ -273,7 +467,7 @@ binning_format = function(binning) {
   ][, bin_iv := lapply(.SD, miv_01, bad), .SDcols = "good"
   ][, total_iv := sum(bin_iv)
   ][, bin := ifelse(is.na(bin) | bin=="NA", "missing", as.character(bin)) # replace NA by missing
-  ][, .(variable, bin, count=good+bad, count_distr=(good+bad)/sum(good+bad), good, bad, badprob, woe, bin_iv, total_iv)]
+  ][, .(variable, bin, count=good+bad, count_distr=(good+bad)/sum(good+bad), good, bad, badprob, woe, bin_iv, total_iv,  breaks = sub("^\\[(.*), *(.*)\\)((%,%missing)*)", "\\2\\3", bin), is_special_values=is_sv)]
 
   # move missing from last row to first
   if ( "missing" %in% binning$bin ) {
@@ -285,60 +479,65 @@ binning_format = function(binning) {
 
 # woebin2
 # This function provides woe binning for only two columns (one x and one y) dataframe.
-woebin2 = function(y, x, x_name, breaks=NULL, min_perc_total=0.02, stop_limit=0.1, max_bin_num=6) {
+woebin2 = function(y, x, x_name, breaks=NULL, spl_val=NULL,  min_perc_fine_bin=0.02, min_perc_coarse_bin=0.05, stop_limit=0.1, max_num_bin=8, method="tree") {
   # global variables or functions
-  . = bad = badprob = bin = bin_iv = good = total_iv = variable = woe = NULL
-
-  # data(germancredit)
-  # numerical data
-  # dt = setDT(germancredit)[,`:=`(y=ifelse(creditability=="good",1,0), creditability=NULL)];
-  # y="y"; x="foreign.worker"age.in.years"
-
-  # categorical data
-  # y="y"; x="status.of.existing.checking.account"
+  . = bad = badprob = bin = bin_iv = good = total_iv = variable = woe = is_sv = NULL
 
   # melt data.table
-  # dtm = data.table(y=dt[[y]], variable=x, value=dt[[x]])
   dtm = data.table(y=y, variable=x_name, value=x)
-
 
   # binning
   if (!anyNA(breaks) & !is.null(breaks)) {
-
     # 1.return binning if breaks provided
-    binning = woebin2_breaks(dtm, breaks)
+    bin_list = woebin2_breaks(dtm=dtm, breaks=breaks, spl_val=spl_val)
 
   } else {
-
-    # 2.tree-like best breaks
-    initial_binning = woebin2_init_bin(dtm, min_perc_total)
     if (stop_limit == "N") {
-      binning = initial_binning
-    } else {
-      binning = woebin2_tree(dtm, initial_binning, min_perc_total, stop_limit, max_bin_num)
-    }
+      # binning of initial & specialvalues
+      bin_list = woebin2_init_bin(dtm, min_perc_fine_bin=min_perc_fine_bin, breaks=breaks, spl_val=spl_val)
 
+    } else {
+      if (method == "tree") {
+        # 2.tree-like optimal binning
+        bin_list = woebin2_tree(dtm, min_perc_fine_bin, min_perc_coarse_bin, stop_limit, max_num_bin, breaks=breaks, spl_val=spl_val)
+
+      } else if (method == "chimerge") {
+        # 2.chimerge optimal binning
+        bin_list = woebin2_chimerge(dtm, min_perc_fine_bin, min_perc_coarse_bin, stop_limit, max_num_bin, breaks=breaks, spl_val=spl_val)
+      }
+    }
   }
+
+  # # binding binning_sv and binning
+  if (any(sapply(bin_list, is.null))) {
+    binning = rbindlist(bin_list)[, is_sv := names(bin_list)[!sapply(bin_list, is.null)]]
+  } else {
+    binning = rbindlist(bin_list, use.names = TRUE, fill = TRUE, idcol = 'is_sv')
+  }
+  binning = binning[, is_sv := is_sv == 'binning_sv']
 
   return(binning_format(binning))
 }
 
 #' WOE Binning
 #'
-#' \code{woebin} generates optimal binning using tree-like segmentation for numerical, factor and categorical variables. \code{woebin} can also customizing breakpoints if the breaks_list was provided.
+#' \code{woebin} generates optimal binning for numerical, factor and categorical variables using methods including tree-like segmentation or chi-square merge. \code{woebin} can also customizing breakpoints if the breaks_list was provided.
 #'
 #' @name woebin
 #' @param dt A data frame with both x (predictor/feature) and y (response/label) variables.
 #' @param y Name of y variable.
 #' @param x Name of x variables. Default is NULL. If x is NULL, then all variables except y are counted as x variables.
-#' @param breaks_list List of break points, defaults NULL If it is not NULL, variable binning will based on the provided breaks.
-#' @param min_perc_total The share of initial binning class number over total. Accepted range: 0.01-0.2; default 0.02.
-#' @param stop_limit Stop binning segmentation when information value gain ratio less than the stop_limit. Accepted range: 0-0.5; default 0.1.
-#' @param max_bin_num Integer. The maximum binning number.
+#' @param breaks_list List of break points, default is NULL. If it is not NULL, variable binning will based on the provided breaks.
+#' @param special_values the values specified in special_values will be in separate bins. Default is NULL.
+#' @param min_perc_fine_bin The minimum percentage of initial binning class number over total. Accepted range: 0.01-0.2; default is 0.02, which means initial binning into 50 fine bins for continuous variables.
+#' @param min_perc_coarse_bin The minimum percentage of final binning class number over total. Accepted range: 0.01-0.2; default is 0.05.
+#' @param stop_limit Stop binning segmentation when information value gain ratio less than the stop_limit, or stop binning merge when the minimum of chi-square less than 'qchisq(1-stoplimit, 1)'. Accepted range: 0-0.5; default is 0.1.
+#' @param max_num_bin Integer. The maximum number of binning.
 #' @param positive Value of positive class, default "bad|1".
-#' @param no_cores Number of CPU cores for parallel computation. Defaults NULL. If no_cores is NULL, the no_cores will set as 1 if length of x variables less than 20, and will set as the number of all CPU cores if the length of x variables greater than or equal to 20.
+#' @param no_cores Number of CPU cores for parallel computation. Defaults NULL. If no_cores is NULL, the no_cores will set as 1 if length of x variables less than 10, and will set as the number of all CPU cores if the length of x variables greater than or equal to 10.
 #' @param print_step A non-negative integer. Default is 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0 or no_cores>1, no message is print.
-#' @return Optimal or customized binning information
+#' @param method Optimal binning method, it should be "tree" or "chimerge". Default is "tree".
+#' @return Optimal or customized binning information.
 #'
 #' @seealso \code{\link{woebin_ply}}, \code{\link{woebin_plot}}, \code{\link{woebin_adj}}
 #'
@@ -359,16 +558,26 @@ woebin2 = function(y, x, x_name, breaks=NULL, min_perc_total=0.02, stop_limit=0.
 #'
 #' # Example III
 #' # customizing the breakpoints of binning
+#' library(data.table)
+#' dat = rbind(
+#'   germancredit,
+#'   data.table(creditability=sample(c("good","bad"),10,replace=TRUE)),
+#'   fill=TRUE)
+#'
 #' breaks_list = list(
-#'   age.in.years = c(26, 35, 37),
-#'   credit.amount = NULL,
-#'   housing = c("own", "for free%,%rent"),
-#'   purpose = NULL
+#'   age.in.years = c(26, 35, 37, "Inf%,%missing"),
+#'   housing = c("own", "for free%,%rent")
 #' )
 #'
-#' bins_cus_brk = woebin(germancredit, y="creditability",
-#'   x=c("age.in.years", "credit.amount", "housing", "purpose"),
-#'   breaks_list=breaks_list)
+#' special_values = list(
+#'   credit.amount = c(2600, 9960, "6850%,%missing"),
+#'   purpose = c("education", "others%,%missing")
+#' )
+#'
+#' bins_cus_brk = woebin(dat, y="creditability",
+#'   x=c("age.in.years","credit.amount","housing","purpose"),
+#'   breaks_list=breaks_list, special_values=special_values)
+#'
 #' }
 #'
 #' @import data.table foreach
@@ -377,7 +586,7 @@ woebin2 = function(y, x, x_name, breaks=NULL, min_perc_total=0.02, stop_limit=0.
 #' @importFrom parallel detectCores
 #' @export
 #'
-woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_limit=0.1, max_bin_num=6, positive="bad|1", no_cores=NULL, print_step=0L) {
+woebin = function(dt, y, x=NULL, breaks_list=NULL, special_values=NULL, min_perc_fine_bin=0.02, min_perc_coarse_bin=0.05, stop_limit=0.1, max_num_bin=8, positive="bad|1", no_cores=NULL, print_step=0L, method="tree") {
   # global variable
   i = NULL
 
@@ -394,31 +603,12 @@ woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_lim
   xs_len = length(xs)
   # print_step
   print_step = check_print_step(print_step)
-
   # breaks_list
-  if (!is.null(breaks_list)) {
-    if (is.character(breaks_list)) {
-      breaks_list = eval(parse(text = breaks_list))
-    }
-    if (!is.list(breaks_list)) {
-      stop("Incorrect inputs; breaks_list should be a list.")
-    } else {
+  breaks_list = check_breaks_list(breaks_list, xs)
+  # special_values
+  special_values = check_special_values(special_values, xs)
 
-      xs_breakslist = names(breaks_list)
-      if (!identical(xs_breakslist, xs)) {
 
-        names_bl_x = setdiff(xs_breakslist, xs)
-        if (length(names_bl_x) > 0) {
-          warning(paste0("Incorrect inputs; the variables \n", paste0(names_bl_x, collapse = ","), "\n specified in breaks_list donot exist in x."))
-        }
-
-        names_x_bl = setdiff(xs, xs_breakslist)
-        if (length(names_x_bl) >0) {
-          warning("There are ",length(names_x_bl)," x variables that donot specified in breaks_list were set as NULL, which means optimal binning.")
-        }
-      }
-    }
-  }
 
   # # stop_limit vector
   # if (length(stop_limit) == 1) {
@@ -432,22 +622,33 @@ woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_lim
     stop_limit = 0.1
   }
 
-  # min_perc_total range
-  if ( min_perc_total<0.01 || min_perc_total>0.2 || !is.numeric(min_perc_total) ) {
-    warning("Incorrect parameter specification; accepted min_perc_total parameter range is 0.01-0.2. Parameter was set to default (0.02).")
-    min_perc_total = 0.02
+  # min_perc_fine_bin range
+  if ( min_perc_fine_bin<0.01 || min_perc_fine_bin>0.2 || !is.numeric(min_perc_fine_bin) ) {
+    warning("Incorrect parameter specification; accepted min_perc_fine_bin parameter range is 0.01-0.2. Parameter was set to default (0.02).")
+    min_perc_fine_bin = 0.02
+  }
+  # min_perc_coarse_bin
+  if ( min_perc_coarse_bin<0.01 || min_perc_coarse_bin>0.2 || !is.numeric(min_perc_coarse_bin) ) {
+    warning("Incorrect parameter specification; accepted min_perc_coarse_bin parameter range is 0.01-0.2. Parameter was set to default (0.05).")
+    min_perc_coarse_bin = 0.05
   }
 
-  # max_bin_num
-  if (!is.numeric(max_bin_num)) {
-    stop("Incorrect inputs; max_bin_num should be numeric variable.")
+  # max_num_bin
+  if (!is.numeric(max_num_bin)) {
+    warning("Incorrect inputs; max_num_bin should be numeric variable. Parameter was set to default (8).")
+    max_num_bin = 8
   }
 
+  # method
+  if (!(method %in% c("tree", "chimerge"))) {
+    warning("Incorrect inputs; method should be tree or chimerge. Parameter was set to default (tree).")
+    method = "tree"
+  }
 
   # binning for each x variable
   # loop on xs # https://www.r-bloggers.com/how-to-go-parallel-in-r-basics-tips/
-  if (is.null(no_cores)) {
-    no_cores = ifelse(xs_len < 20, 1, detectCores())
+  if (is.null(no_cores) || no_cores<1) {
+    no_cores = ifelse(xs_len < 10, 1, detectCores())
   }
 
 
@@ -464,8 +665,11 @@ woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_lim
         woebin2(
           y=dt[[y]], x=dt[[x_i]], x_name=x_i,
           breaks=breaks_list[[x_i]],
-          min_perc_total=min_perc_total,
-          stop_limit=stop_limit, max_bin_num=max_bin_num
+          spl_val=special_values[[x_i]],
+          min_perc_fine_bin=min_perc_fine_bin,
+          min_perc_coarse_bin=min_perc_coarse_bin,
+          stop_limit=stop_limit, max_num_bin=max_num_bin,
+          method=method
         ),
         error = function(e) return(paste0("The variable '", x_i, "'", " caused the error: '", e, "'"))
       )
@@ -486,7 +690,7 @@ woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_lim
           if (xs_len==1) bs = list(bs)
           setNames(bs, xs)
         },
-        .export = c("dt", "xs", "y", "breaks_list", "min_perc_total", "stop_limit", "max_bin_num")
+        .export = c("dt", "xs", "y", "breaks_list", "min_perc_fine_bin", "stop_limit", "max_num_bin")
       ) %dopar% {
         x_i = xs[i]
 
@@ -494,8 +698,12 @@ woebin = function(dt, y, x=NULL, breaks_list=NULL, min_perc_total=0.02, stop_lim
         tryCatch(
           woebin2(
             y=dt[[y]], x=dt[[x_i]], x_name=x_i,
-            breaks=breaks_list[[x_i]], min_perc_total=min_perc_total,
-            stop_limit=stop_limit, max_bin_num=max_bin_num
+            breaks=breaks_list[[x_i]],
+            spl_val=special_values[[x_i]],
+            min_perc_fine_bin=min_perc_fine_bin,
+            min_perc_coarse_bin=min_perc_coarse_bin,
+            stop_limit=stop_limit, max_num_bin=max_num_bin,
+            method=method
           ),
           error = function(e) return(paste0("The variable '", x_i, "'", " caused the error: '", e, "'"))
         )
@@ -525,7 +733,15 @@ woepoints_ply1 = function(dtx, binx, x_i, woe_points) {
   # dtx
   ## cut numeric variable
   if ( is.numeric(dtx[[x_i]]) ) {
-    dtx[[x_i]] = cut(dtx[[x_i]], unique(c(-Inf, binx[bin != "missing", as.numeric(sub("^\\[(.*),.+", "\\1", bin))], Inf)), right = FALSE, dig.lab = 10, ordered_result = FALSE)
+    binx_sv = binx[!grepl("\\[",V1)]
+    binx_other = binx[grepl("\\[",V1)]
+
+    dtx[[x_i]] = ifelse(
+      dtx[[x_i]] %in% binx_sv$V1,
+      dtx[[x_i]],
+      as.character(cut(dtx[[x_i]], unique(c(-Inf, binx_other[bin != "missing", as.numeric(sub("^\\[(.*),.+", "\\1", V1))], Inf)), right = FALSE, dig.lab = 10, ordered_result = FALSE))
+    )
+
   }
   ## to charcarter, na to missing
   dtx[[x_i]] = as.character(dtx[[x_i]])
@@ -547,7 +763,7 @@ woepoints_ply1 = function(dtx, binx, x_i, woe_points) {
 #'
 #' @param dt A data frame.
 #' @param bins Binning information generated from \code{woebin}.
-#' @param no_cores Number of CPU cores for parallel computation. Defaults NULL. If no_cores is NULL, the no_cores will set as 1 if length of x variables less than 20, and will set as the number of all CPU cores if the length of x variables greater than or equal to 20.
+#' @param no_cores Number of CPU cores for parallel computation. Defaults NULL. If no_cores is NULL, the no_cores will set as 1 if length of x variables less than 10, and will set as the number of all CPU cores if the length of x variables greater than or equal to 10.
 #' @param print_step A non-negative integer. Default is 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0 or no_cores>1, no message is print.
 #' @return Binning information
 #'
@@ -594,7 +810,7 @@ woebin_ply = function(dt, bins, no_cores=NULL, print_step=0L) {
   # replace "" by NA
   dt = rep_blank_na(dt)
   # ncol of dt
-  if (ncol(dt) <=1 & !is.null(ncol(dt))) stop("Incorrect inputs; dt should have at least two columns.")
+  # if (ncol(dt) <=1 & !is.null(ncol(dt))) stop("Incorrect inputs; dt should have at least two columns.")
   # print_step
   print_step = check_print_step(print_step)
 
@@ -620,7 +836,7 @@ woebin_ply = function(dt, bins, no_cores=NULL, print_step=0L) {
 
   # loop on xs # https://www.r-bloggers.com/how-to-go-parallel-in-r-basics-tips/
   if (is.null(no_cores)) {
-    no_cores = ifelse(xs_len < 20, 1, detectCores())
+    no_cores = ifelse(xs_len < 10, 1, detectCores())
   }
 
   if (no_cores == 1) {
@@ -792,7 +1008,9 @@ woebin_plot = function(bins, x=NULL, title=NULL, show_iv = TRUE) {
 
 
 # print basic information in woebin_adj
-woebin_adj_print = function(i, xs_len, x_i, bins, dt, bins_breakslist) {
+woebin_adj_print_basic_info = function(i, xs_adj, bins, dt, bins_breakslist) {
+  x_i = xs_adj[i]
+  xs_len = length(xs_adj)
   variable = x_breaks = NULL
 
   bin = bins[variable==x_i]
@@ -805,7 +1023,7 @@ woebin_adj_print = function(i, xs_len, x_i, bins, dt, bins_breakslist) {
   print(summary(dt[[x_i]]))
   cat("\n")
   ## table
-  if (length(table(dt[[x_i]])) < 10) {
+  if (length(table(dt[[x_i]])) < 10 || !is.numeric(dt[[x_i]])) {
     cat(paste0("> table(",x_i,"): "))
     print(table(dt[[x_i]]))
     cat("\n")
@@ -825,16 +1043,16 @@ woebin_adj_print = function(i, xs_len, x_i, bins, dt, bins_breakslist) {
 
 }
 # plot adjusted binning in woebin_adj
-woebin_adj_break_plot = function(dt, y, x_i, breaks, stop_limit) {
+woebin_adj_break_plot = function(dt, y, x_i, breaks, stop_limit, special_values, method) {
   bin_adj = NULL
 
-  text_woebin = paste0("bin_adj=woebin(dt[,c(\"",x_i,"\",\"",y,"\"),with=F], \"",y,"\", breaks_list=list(",x_i,"=c(",breaks,")), ", ifelse(stop_limit=="N","stop_limit = \"N\", ",NULL), "print_step=0L)")
+  text_woebin = paste0("bin_adj=woebin(dt[,c(\"",x_i,"\",\"",y,"\"),with=F], \"",y,"\", breaks_list=list(",x_i,"=c(",breaks,")), special_values =", special_values, ", ", ifelse(stop_limit=="N","stop_limit = \"N\", ",NULL), "print_step=0L, method=\"",method,"\")")
 
   eval(parse(text = text_woebin))
 
 
   ## print adjust breaks
-  breaks_bin = setdiff(sub("^\\[(.*),.+", "\\1", bin_adj[[1]]$bin), c("-Inf","Inf","missing"))
+  breaks_bin = setdiff(sub("^\\[(.*), *(.*)\\)((%,%missing)*)", "\\2\\3", bin_adj[[1]]$bin), c("-Inf","Inf","missing"))
   breaks_bin = ifelse(
     is.numeric(dt[[x_i]]),
     paste0(breaks_bin, collapse=", "),
@@ -856,8 +1074,9 @@ woebin_adj_break_plot = function(dt, y, x_i, breaks, stop_limit) {
 #' @param dt A data frame.
 #' @param y Name of y variable.
 #' @param bins A list or data frame. Binning information generated from \code{woebin}.
-#' @param all_var Logical, default is TRUE. If it is TRUE, the variables that need to adjust breaks include all variables, otherwise, include the variables that number distribution rate less then count_distr_limit or more then one inflection point.
-#' @param count_distr_limit The count_distr limit to adjust binning breaks. Default 0.05.
+#' @param adj_all_var Logical, default is TRUE. If it is TRUE, all variables need to adjust binning breaks, otherwise, only include the variables that have more then one inflection point.
+#' @param special_values the values specified in special_values will in separate bins. Default is NULL.
+#' @param method optimal binning method, it should be "tree" or "chimerge". Default is "tree".
 #'
 #' @seealso  \code{\link{woebin}}, \code{\link{woebin_ply}}, \code{\link{woebin_plot}}
 #'
@@ -885,7 +1104,7 @@ woebin_adj_break_plot = function(dt, y, x_i, breaks, stop_limit) {
 #' @importFrom graphics hist plot
 #' @export
 #'
-woebin_adj = function(dt, y, bins, all_var = TRUE, count_distr_limit = 0.05) {
+woebin_adj = function(dt, y, bins, adj_all_var=TRUE, special_values=NULL, method="tree") {
   # global variables or functions
   . = V1 = badprob = badprob2 = bin2 = bin = bin_adj = count_distr = variable = x_breaks = x_class = NULL
 
@@ -901,11 +1120,12 @@ woebin_adj = function(dt, y, bins, all_var = TRUE, count_distr_limit = 0.05) {
 
   # x variables
   xs_all = bins[,unique(variable)]
-  if (all_var == FALSE) {
-    xs_adj = union(
-      bins[bin != "missing"][count_distr < count_distr_limit, unique(variable)],
-      bins[bin != "missing"][, badprob2 := badprob >= shift(badprob, type = "lag"), by="variable"][!is.na(badprob2), length(unique(badprob2)), by="variable"][V1 > 1, variable]
-    )
+  if (adj_all_var == FALSE) {
+    xs_adj = bins[
+      bin != "missing"
+    ][, badprob2 := badprob >= shift(badprob, type = "lag"), by=variable
+    ][!is.na(badprob2), length(unique(badprob2)), by=variable
+    ][V1 > 1, variable]
   } else {
     xs_adj = xs_all
   }
@@ -919,7 +1139,7 @@ woebin_adj = function(dt, y, bins, all_var = TRUE, count_distr_limit = 0.05) {
   )
   # breakslist of bins
   bins_breakslist = bins[
-    , bin2 := sub("^\\[(.*),.+", "\\1", bin)
+    , bin2 := sub("^\\[(.*), *(.*)\\)((%,%missing)*)", "\\2\\3", bin)
     ][!(bin2 %in% c("-Inf","Inf","missing"))
     ][vars_class, on="variable"
     ][, .(
@@ -946,7 +1166,7 @@ woebin_adj = function(dt, y, bins, all_var = TRUE, count_distr_limit = 0.05) {
     x_i = xs_adj[i]
 
     # basic information of x_i variable ------
-    woebin_adj_print(i, xs_len, x_i, bins, dt, bins_breakslist)
+    woebin_adj_print_basic_info(i, xs_adj, bins, dt, bins_breakslist)
 
     # adjusting breaks ------
     adj_brk = menu(c("next", "yes", "back"), title=paste0("> Adjust breaks for (", i, "/", xs_len, ") ", x_i, "?"))
@@ -962,7 +1182,7 @@ woebin_adj = function(dt, y, bins, all_var = TRUE, count_distr_limit = 0.05) {
         stop_limit = NULL
       }
 
-      tryCatch(breaks <- woebin_adj_break_plot(dt, y, x_i, breaks, stop_limit), error = function(e) e)
+      tryCatch(breaks <- woebin_adj_break_plot(dt, y, x_i, breaks, stop_limit, special_values, method=method), error = function(e) e)
 
       adj_brk = menu(c("next", "yes", "back"), title=paste0("> Adjust breaks for (", i, "/", xs_len, ") ", x_i, "?"))
     }