+++ NEWS.md
@@ -1,3 +1,9 @@
+# scorecard 0.1.7
+
+* added chimerge method for woebin function
+* special_values option added in woebin function
+* f1 curve added in perf_eva
+
 # scorecard 0.1.6
 
 * Fixed a bug in woebin_adj function when all_var == FALSE and the breaks of all variables are perfect.